Refactor BehaviourTestBase to separate wholesale services and aggregated measure data
