+++ source/IntegrationTests/Behaviours/BehavioursTestBase.cs
@@ -295,150 +295,12 @@ public class BehavioursTestBase : IDisposable
                 CancellationToken.None);
     }
 
-    protected async Task<ResponseMessage> GivenReceivedAggregatedMeasureDataRequest(
-        DocumentFormat documentFormat,
-        ActorNumber senderActorNumber,
-        ActorRole senderActorRole,
-        MeteringPointType? meteringPointType,
-        SettlementMethod? settlementMethod,
-        (int Year, int Month, int Day) periodStart,
-        (int Year, int Month, int Day) periodEnd,
-        ActorNumber? energySupplier,
-        ActorNumber? balanceResponsibleParty,
-        IReadOnlyCollection<(string? GridArea, string TransactionId)> series,
-        bool assertRequestWasSuccessful = true)
-    {
-        var incomingMessageClient = GetService<IIncomingMessageClient>();
-
-        var incomingMessageStream = RequestAggregatedMeasureDataRequestBuilder.CreateIncomingMessage(
-            format: documentFormat,
-            senderActorNumber: senderActorNumber,
-            senderActorRole: senderActorRole,
-            meteringPointType: meteringPointType,
-            settlementMethod: settlementMethod,
-            periodStart: CreateDateInstant(periodStart.Year, periodStart.Month, periodStart.Day),
-            periodEnd: CreateDateInstant(periodEnd.Year, periodEnd.Month, periodEnd.Day),
-            energySupplier: energySupplier,
-            balanceResponsibleParty: balanceResponsibleParty,
-            series: series,
-            ensureValidRequest: assertRequestWasSuccessful);
-
-        var response = await
-            incomingMessageClient.RegisterAndSendAsync(
-                incomingMessageStream,
-                documentFormat,
-                IncomingDocumentType.RequestAggregatedMeasureData,
-                documentFormat,
-                CancellationToken.None);
-
-        if (assertRequestWasSuccessful)
-        {
-            using var scope = new AssertionScope();
-            response.IsErrorResponse.Should().BeFalse("because the response should not have an error. Actual response: {0}", response.MessageBody);
-            response.MessageBody.Should().BeEmpty();
-        }
-
-        return response;
-    }
-
-    protected async Task<ResponseMessage> GivenReceivedWholesaleServicesRequest(
-        DocumentFormat documentFormat,
-        ActorNumber senderActorNumber,
-        ActorRole senderActorRole,
-        (int Year, int Month, int Day) periodStart,
-        (int Year, int Month, int Day) periodEnd,
-        ActorNumber? energySupplierActorNumber,
-        ActorNumber? chargeOwnerActorNumber,
-        string? chargeCode,
-        ChargeType? chargeType,
-        bool isMonthly,
-        IReadOnlyCollection<(string? GridArea, string TransactionId)> series,
-        bool assertRequestWasSuccessful = true)
-    {
-        var incomingMessageClient = GetService<IIncomingMessageClient>();
-
-        var incomingMessageStream = RequestWholesaleServicesRequestBuilder.GetStream(
-            documentFormat,
-            senderActorNumber,
-            senderActorRole,
-            CreateDateInstant(periodStart.Year, periodStart.Month, periodStart.Day),
-            CreateDateInstant(periodEnd.Year, periodEnd.Month, periodEnd.Day),
-            energySupplierActorNumber,
-            chargeOwnerActorNumber,
-            chargeCode,
-            chargeType,
-            isMonthly,
-            series);
-
-        var response = await
-            incomingMessageClient.RegisterAndSendAsync(
-                incomingMessageStream,
-                documentFormat,
-                IncomingDocumentType.RequestWholesaleSettlement,
-                documentFormat,
-                CancellationToken.None);
-
-        if (assertRequestWasSuccessful)
-        {
-            using var scope = new AssertionScope();
-            response.IsErrorResponse.Should().BeFalse();
-            response.MessageBody.Should().BeEmpty();
-        }
-
-        return response;
-    }
-
-    protected async Task GivenInitializeAggregatedMeasureDataProcessDtoIsHandledAsync(
-        ServiceBusMessage serviceBusMessage)
-    {
-        // We have to manually process the service bus message, as there isn't a real service bus
-        serviceBusMessage.Subject.Should().Be(nameof(InitializeAggregatedMeasureDataProcessDto));
-        serviceBusMessage.Body.Should().NotBeNull();
-
-        await GetService<IProcessClient>().InitializeAsync(serviceBusMessage.Subject, serviceBusMessage.Body.ToArray());
-        await ProcessInternalCommandsAsync();
-    }
-
-    protected async Task WhenInitializeAggregatedMeasureDataProcessDtoIsHandledAsync(
-        ServiceBusMessage serviceBusMessage)
-    {
-        await GivenInitializeAggregatedMeasureDataProcessDtoIsHandledAsync(serviceBusMessage);
-    }
-
-    protected async Task GivenWholesaleAcceptedResponseToAggregatedMeasureDataRequestAsync(
-        ServiceBusMessage serviceBusMessage)
-    {
-        serviceBusMessage.Subject.Should().Be(nameof(AggregatedTimeSeriesRequest));
-        serviceBusMessage.Body.Should().NotBeNull();
-
-        var aggregatedTimeSeriesRequest =
-            AggregatedTimeSeriesRequest.Parser.ParseFrom(serviceBusMessage.Body);
-
-        var aggregatedTimeSeriesRequestAccepted =
-            AggregatedTimeSeriesResponseEventBuilder.GenerateAcceptedFrom(aggregatedTimeSeriesRequest, GetNow());
-
-        await HavingReceivedInboxEventAsync(
-            nameof(AggregatedTimeSeriesRequestAccepted),
-            aggregatedTimeSeriesRequestAccepted,
-            Guid.Parse(serviceBusMessage.MessageId));
-    }
-
     protected ServiceBusSenderSpy CreateServiceBusSenderSpy()
     {
         var serviceBusSenderSpy = new ServiceBusSenderSpy(MockServiceBusName);
         _serviceBusSenderFactoryStub.AddSenderSpy(serviceBusSenderSpy);
 
         return serviceBusSenderSpy;
-    }
-
-    protected async Task WhenWholesaleServicesProcessIsInitialized(ServiceBusMessage serviceBusMessage)
-    {
-        await InitializeProcess(serviceBusMessage, nameof(InitializeWholesaleServicesProcessDto));
-    }
-
-    protected async Task WhenAggregatedMeasureDataProcessIsInitialized(ServiceBusMessage serviceBusMessage)
-    {
-        await InitializeProcess(serviceBusMessage, nameof(InitializeAggregatedMeasureDataProcessDto));
     }
 
     protected (TServiceBusMessage Message, Guid ProcessId) AssertServiceBusMessage<TServiceBusMessage>(ServiceBusSenderSpy senderSpy, Func<BinaryData, TServiceBusMessage> parser)
@@ -502,67 +364,6 @@ public class BehavioursTestBase : IDisposable
         return peekResults;
     }
 
-    protected async Task ThenNotifyWholesaleServicesDocumentIsCorrect(
-        Stream? peekResultDocumentStream,
-        DocumentFormat documentFormat,
-        NotifyWholesaleServicesDocumentAssertionInput assertionInput)
-    {
-        peekResultDocumentStream.Should().NotBeNull();
-        peekResultDocumentStream!.Position = 0;
-
-        using var assertionScope = new AssertionScope();
-
-        await NotifyWholesaleServicesDocumentAsserter.AssertCorrectDocumentAsync(
-            documentFormat,
-            peekResultDocumentStream,
-            assertionInput);
-    }
-
-    protected async Task ThenNotifyAggregatedMeasureDataDocumentIsCorrect(
-        Stream? peekResultDocumentStream,
-        DocumentFormat documentFormat,
-        NotifyAggregatedMeasureDataDocumentAssertionInput assertionInput)
-    {
-        peekResultDocumentStream.Should().NotBeNull();
-        peekResultDocumentStream!.Position = 0;
-
-        using var assertionScope = new AssertionScope();
-
-        await NotifyAggregatedMeasureDataDocumentAsserter.AssertCorrectDocumentAsync(
-            documentFormat,
-            peekResultDocumentStream,
-            assertionInput);
-    }
-
-    protected async Task ThenRejectRequestWholesaleSettlementDocumentIsCorrect(Stream? peekResultDocumentStream, DocumentFormat documentFormat, RejectRequestWholesaleSettlementDocumentAssertionInput assertionInput)
-    {
-        peekResultDocumentStream.Should().NotBeNull();
-        peekResultDocumentStream!.Position = 0;
-
-        using var assertionScope = new AssertionScope();
-
-        await RejectRequestWholesaleSettlementDocumentAsserter.AssertCorrectDocumentAsync(
-            documentFormat,
-            peekResultDocumentStream,
-            assertionInput);
-    }
-
-    protected async Task ThenRejectRequestAggregatedMeasureDataDocumentIsCorrect(
-        Stream? peekResultDocumentStream,
-        DocumentFormat documentFormat,
-        RejectRequestAggregatedMeasureDataDocumentAssertionInput assertionInput)
-    {
-        peekResultDocumentStream.Should().NotBeNull();
-        peekResultDocumentStream!.Position = 0;
-
-        using var assertionScope = new AssertionScope();
-
-        await RejectRequestAggregatedMeasureDataDocumentAsserter.AssertCorrectDocumentAsync(
-            documentFormat,
-            peekResultDocumentStream,
-            assertionInput);
-    }
-
     protected AmountPerChargeResultProducedV1 GivenAmountPerChargeResultProducedV1Event(Action<AmountPerChargeResultProducedV1EventBuilder> builder)
     {
         var eventBuilder = new AmountPerChargeResultProducedV1EventBuilder();
@@ -572,113 +373,13 @@ public class BehavioursTestBase : IDisposable
         return eventBuilder.Build();
     }
 
-    protected Task GivenAggregatedMeasureDataRequestAcceptedIsReceived(Guid processId, AggregatedTimeSeriesRequestAccepted acceptedMessage)
+    protected T GetService<T>()
+        where T : notnull
     {
-        return HavingReceivedInboxEventAsync(
-            eventType: nameof(AggregatedTimeSeriesRequestAccepted),
-            eventPayload: acceptedMessage,
-            processId: processId);
+        return _serviceProvider.GetRequiredService<T>();
     }
 
-    protected async Task<string> GetGridAreaFromNotifyAggregatedMeasureDataDocument(Stream documentStream, DocumentFormat documentFormat)
-    {
-        documentStream.Position = 0;
-        if (documentFormat == DocumentFormat.Ebix)
-        {
-            var ebixAsserter = NotifyAggregatedMeasureDataDocumentAsserter.CreateEbixAsserter(documentStream);
-            var gridAreaElement = ebixAsserter.GetElement("PayloadEnergyTimeSeries[1]/MeteringGridAreaUsedDomainLocation/Identification");
-
-            gridAreaElement.Should().NotBeNull("because grid area should be present in the ebIX document");
-            gridAreaElement!.Value.Should().NotBeNull("because grid area value should not be null in the ebIX document");
-            return gridAreaElement.Value;
-        }
-
-        if (documentFormat == DocumentFormat.Xml)
-        {
-            var cimXmlAsserter = NotifyAggregatedMeasureDataDocumentAsserter.CreateCimXmlAsserter(documentStream);
-
-            var gridAreaCimXmlElement = cimXmlAsserter.GetElement("Series[1]/meteringGridArea_Domain.mRID");
-
-            gridAreaCimXmlElement.Should().NotBeNull("because grid area should be present in the CIM XML document");
-            gridAreaCimXmlElement!.Value.Should().NotBeNull("because grid area value should not be null in the CIM XML document");
-            return gridAreaCimXmlElement!.Value;
-        }
-
-        if (documentFormat == DocumentFormat.Json)
-        {
-            var cimJsonDocument = await JsonDocument.ParseAsync(documentStream);
-
-            var gridAreaCimJsonElement = cimJsonDocument.RootElement
-                .GetProperty("NotifyAggregatedMeasureData_MarketDocument")
-                .GetProperty("Series").EnumerateArray().ToList()
-                .Single()
-                .GetProperty("meteringGridArea_Domain.mRID")
-                .GetProperty("value");
-
-            gridAreaCimJsonElement.Should().NotBeNull("because grid area should be present in the CIM JSON document");
-            return gridAreaCimJsonElement.GetString()!;
-        }
-
-        throw new ArgumentOutOfRangeException(nameof(documentFormat), documentFormat, "Unsupported document format");
-    }
-
-    protected Task<(AggregatedTimeSeriesRequest AggregatedTimeSeriesRequest, Guid ProcessId)> ThenAggregatedTimeSeriesRequestServiceBusMessageIsCorrect(
-            ServiceBusSenderSpy senderSpy,
-            IReadOnlyCollection<string> gridAreas,
-            string requestedForActorNumber,
-            string requestedForActorRole,
-            string? energySupplier,
-            string? balanceResponsibleParty,
-            BusinessReason businessReason,
-            Period period,
-            SettlementVersion? settlementVersion,
-            SettlementMethod? settlementMethod,
-            MeteringPointType? meteringPointType)
-    {
-        var (message, processId) = AssertServiceBusMessage(
-            senderSpy,
-            data => AggregatedTimeSeriesRequest.Parser.ParseFrom(data));
-
-        using var assertionScope = new AssertionScope();
-
-        message.GridAreaCodes.Should().BeEquivalentTo(gridAreas);
-        message.RequestedForActorNumber.Should().Be(requestedForActorNumber);
-        message.RequestedForActorRole.Should().Be(requestedForActorRole);
-
-        if (energySupplier == null)
-            message.HasEnergySupplierId.Should().BeFalse();
-        else
-            message.EnergySupplierId.Should().Be(energySupplier);
-
-        if (balanceResponsibleParty == null)
-            message.HasBalanceResponsibleId.Should().BeFalse();
-        else
-            message.BalanceResponsibleId.Should().Be(balanceResponsibleParty);
-
-        message.BusinessReason.Should().Be(businessReason.Name);
-
-        message.Period.Start.Should().Be(period.Start.ToString());
-        message.Period.End.Should().Be(period.End.ToString());
-
-        if (settlementVersion == null)
-            message.HasSettlementVersion.Should().BeFalse();
-        else
-            message.SettlementVersion.Should().Be(settlementVersion.Name);
-
-        if (settlementMethod == null)
-            message.HasSettlementMethod.Should().BeFalse();
-        else
-            message.SettlementMethod.Should().Be(settlementMethod.Name);
-
-        if (meteringPointType == null)
-            message.MeteringPointType.Should().BeEmpty(); // Contract is incorrect and doesn't have MeteringPointType as optional
-        else
-            message.MeteringPointType.Should().Be(meteringPointType.Name);
-
-        return Task.FromResult((message, processId));
-    }
-
-    private async Task InitializeProcess(ServiceBusMessage serviceBusMessage, string expectedSubject)
+    protected async Task InitializeProcess(ServiceBusMessage serviceBusMessage, string expectedSubject)
     {
         using var scope = _serviceProvider.CreateScope();
         // We have to manually process the service bus message, as there isn't a real service bus
@@ -689,14 +390,7 @@ public class BehavioursTestBase : IDisposable
         await ProcessInternalCommandsAsync();
     }
 
-    private async Task WhenActorDequeuesMessage(string messageId, ActorNumber actorNumber, ActorRole actorRole)
-    {
-        await using var scope = _serviceProvider.CreateAsyncScope();
-        var outgoingMessagesClient = scope.ServiceProvider.GetRequiredService<IOutgoingMessagesClient>();
-        await outgoingMessagesClient.DequeueAndCommitAsync(new DequeueRequestDto(messageId, actorRole, actorNumber), CancellationToken.None);
-    }
-
-    private async Task ProcessInternalCommandsAsync()
+    protected async Task ProcessInternalCommandsAsync()
     {
         await ProcessBackgroundTasksAsync();
 
@@ -706,10 +400,11 @@ public class BehavioursTestBase : IDisposable
         }
     }
 
-    private T GetService<T>()
-        where T : notnull
+    private async Task WhenActorDequeuesMessage(string messageId, ActorNumber actorNumber, ActorRole actorRole)
     {
-        return _serviceProvider.GetRequiredService<T>();
+        await using var scope = _serviceProvider.CreateAsyncScope();
+        var outgoingMessagesClient = scope.ServiceProvider.GetRequiredService<IOutgoingMessagesClient>();
+        await outgoingMessagesClient.DequeueAndCommitAsync(new DequeueRequestDto(messageId, actorRole, actorNumber), CancellationToken.None);
     }
 
     private Task ProcessReceivedInboxEventsAsync()
